reset unnecessary changes
